requirements.txt
anyio==4.2.0
certifi==2024.2.2
dnspython==2.6.0
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
idna==3.6
pymongo==4.6.1
python-dotenv==1.0.1
python-telegram-bot==20.8
sniffio==1.3.0
